Pc_server_Application
This project is based on socet programming using c++. The PC as a server and the phone as a client. Both(server and client) must be connected to the same network,You can view and edit files, run programs, and perform tasks on your PC from anywhere with an internet connection.